| optim: optimize the draw function by giving some char for x and y | center the title on the main screen and add Rick sprite

+++ CustomArduboy.h
@@ -134,7 +134,7 @@ const uint8_t TRANSPARENT = 3;
 class CustomArduboy : public Arduboy
 {
 public:
-	unsigned int drawBitmapExtended(int16_t x, int16_t y, const uint8_t *bitmap, uint8_t w, uint8_t h, uint8_t color, bool mirrorX);
+	unsigned int drawBitmapExtended(int8_t x, int8_t y, const uint8_t *bitmap, uint8_t w, uint8_t h, uint8_t color, bool mirrorX);
 	bool CheckWhitePixelsInRow(uint8_t x, uint8_t y, uint8_t w);
 	uint8_t GetCursorX() { return cursor_x; }
 	uint8_t GetCursorY() { return cursor_y; }
